# perf: 优化代码速度