提速并按`gofmt`规范化代码格式
```
name     old time/op    new time/op    delta
Av2Bv-8    76.2ns ± 1%    31.7ns ± 0%  -58.47%  (p=0.008 n=5+5)
Bv2Av-8     132ns ± 1%     118ns ± 1%  -10.42%  (p=0.008 n=5+5)
```